Fix some links?
GitHub Pages adds a subdirectory for the repository name which isn't present
when testing locally, so these errors only become apparent after deploying.

This also means that I can't be certain that this commit fixes the issues...

Moreover, some of these links cannot be expressed consistently between the
deployment and local environments without resorting to a mess of ../../../../
etc. Should we migrate to Jekyll to handle this? It does add some complexity
itself...